+++ src/api.js
@@ -1,0 +1,8 @@
+const API_URL = 'https://jsonplaceholder.typicode.com/';
+
+const getData = dataName => (
+  fetch(`${API_URL}${dataName}`)
+    .then(response => response.json())
+);
+
+export default getData;
